compute: move `MaybeValidatingRow` into `errors` module
